asdas

+++ engineeringOffice/src/com/entities/Bills.java
@@ -1,5 +1,6 @@
 package com.entities;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 import javax.persistence.Column;
@@ -30,10 +31,10 @@ public class Bills {
 	private String amountLetters;
 
 	@Column(name = "amount")
-	private Integer amountPay;
+	private BigDecimal amountPay;
 
 	@Column(name = "amount_rest")
-	private Integer amountRest;
+	private BigDecimal amountRest;
 
 	@Column(name = "bill_reason")
 	private String billReason;
@@ -90,19 +91,19 @@ public class Bills {
 		this.amountLetters = amountLetters;
 	}
 
-	public Integer getAmountPay() {
+	public BigDecimal getAmountPay() {
 		return amountPay;
 	}
 
-	public void setAmountPay(Integer amountPay) {
+	public void setAmountPay(BigDecimal amountPay) {
 		this.amountPay = amountPay;
 	}
 
-	public Integer getAmountRest() {
+	public BigDecimal getAmountRest() {
 		return amountRest;
 	}
 
-	public void setAmountRest(Integer amountRest) {
+	public void setAmountRest(BigDecimal amountRest) {
 		this.amountRest = amountRest;
 	}
